[PimcoreBundle] fix dynamic selects with Javascript
